--- SurveyQuestions.json
@@ -1,28 +1,0 @@
-
-{
-  "surveyPatients": [
-    {
-     "question" : "Imię Pacjenta",
-      "answers":[ "Mężczyzna", "Kobieta"
-
-      ]
-    },
-    {
-      "question" : "Nazwisko Pacjenta",
-      "answers":[ "99", "88"
-
-      ]
-    },
-
-    {
-      "question" : "PESEL Pacjenta ",
-      "answers":[ "99", "88"
-
-      ]
-    }
-
-  ]
-}
-
-
-
