--- listaDoktorów.json
@@ -1,19 +1,0 @@
-
-{
-  "doctors": [
-    {
-      "id": 234232,
-      "specjalizacja": "Laryngolog",
-      "name": "Rafa\u0142",
-      "lastName": "Mirkowski"
-    },
-    {
-      "id": 23432223,
-      "specjalizacja": "Pediatra",
-      "name": "Robert",
-      "lastName": "Kowalski"
-    }
-  ]
-}
-
-
