--- mycalendar2.ics
@@ -1,11 +1,0 @@
-BEGIN:VCALENDAR
-PRODID:/Medicine/
-VERSION:2.0
-CALSCALE:GREGORIAN
-BEGIN:VEVENT
-DTSTAMP:20170312T191226Z
-DTSTAMP:20170312T191226Z
-SUMMARY:Appointment surveyResultPatient
-UID:20170312T191226Z-1@fe80:0:0:0:1513:ec94:231d:20d7%eth1
-END:VEVENT
-END:VCALENDAR
