--- mycalendar3.ics
@@ -1,11 +1,0 @@
-BEGIN:VCALENDAR
-PRODID:/Medicine/
-VERSION:2.0
-CALSCALE:GREGORIAN
-BEGIN:VEVENT
-DTSTAMP:20170216T220813Z
-DTSTART;VALUE=DATE;VALUE=DATE:20170214
-SUMMARY:Vallentines Day
-UID:20170216T220813Z-1@fe80:0:0:0:3282:3fba:f545:ad37%wlp4s0
-END:VEVENT
-END:VCALENDAR
