--- src/main/java/Exceptions/WrongOptionsExeption.java
@@ -1,13 +1,0 @@
-package Exceptions;
-
-/**
- * Created by moody on 02.03.17.
- */
-public class WrongOptionsExeption extends RuntimeException {
-
-    public WrongOptionsExeption (String wrongOptionExeption){
-        super(wrongOptionExeption);
-
-    }
-
-}
