--- src/main/java/infoshare/academy/peanut/medicine/Doctor.java
@@ -1,65 +1,0 @@
-package infoshare.academy.peanut.medicine;
-
-import java.lang.reflect.Array;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-/**
- * Created by bartman3000 on 2017-03-11.
- */
-public class Doctor {
-
-    private String name;
-    private String surname;
-    private String specialization;
-    private Set<LocalDate> terms;
-
-    public Doctor(String name, String surname, String specialization) {
-        this.name = name;
-        this.surname = surname;
-        this.specialization = specialization;
-        this.terms = new HashSet<LocalDate>();
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getSpecialization() {
-        return specialization;
-    }
-
-    public void setSpecialization(String specialization) {
-        this.specialization = specialization;
-    }
-
-    public Set<LocalDate> getTerms() {
-        return terms;
-    }
-
-    public void addTerm(LocalDate d)
-    {
-        terms.add(d);
-    }
-
-    public String toString()
-    {
-        return this.specialization + " " + this.name + " "+this.surname + "" + this.terms.toString();
-    }
-}
