--- src/main/java/infoshare/academy/peanut/medicine/iCalendar/IcalendarTest.java
@@ -1,33 +1,0 @@
-package infoshare.academy.peanut.medicine.iCalendar;
-
-import net.fortuna.ical4j.model.Calendar;
-
-import java.io.File;
-
-/**
- * Created by Mariusz on 2017-03-03.
- */
-public class IcalendarTest {
-
-    public static void main(String[] args) {
-
-        // Pathname to .ics calendar file
-        File icsFile = new File("./mycalendar2.ics");
-
-        // Parsing an iCalendar .ics file
-        IcalendarReaderICS icalr = new IcalendarReaderICS();
-        Calendar calendar = icalr.readCalendar(icsFile);
-
-        // Creating a meeting of four hour duration
-        IcalendarMeeting icalendarMeeting = new IcalendarMeeting();
-        Calendar newCalendar = icalendarMeeting.setMeeting(calendar);
-
-        // Showing calendar file
-        ShowFileICS showFileICS = new ShowFileICS();
-        showFileICS.printIcsData(calendar);
-
-        // Saving calendar -> .ics file
-        IcalendarWriterICS icalendarWriterICS = new IcalendarWriterICS();
-        icalendarWriterICS.writeCalendar(newCalendar,icsFile);
-    }
-}
