--- src/main/java/infoshare/academy/peanut/medicine/survey/SurveyQuestionList.java
@@ -1,17 +1,0 @@
-package infoshare.academy.peanut.medicine.survey;
-
-/**
- * Created by moody on 20.02.17.
- */
-public class SurveyQuestionList {
-
-   private SurveyPatient[] surveyPatients ;
-
-    public SurveyPatient[] getSurveyPatients(){
-        return surveyPatients ;
-    }
-
-    public void setSurveyPatients(SurveyPatient[] surveyPatients) {
-        this.surveyPatients = surveyPatients;
-    }
-}
