+++ src/main/java/peanut/medicine/Exceptions/WrongOptionsExeption.java
@@ -1,0 +1,13 @@
+package peanut.medicine.Exceptions;
+
+/**
+ * Created by moody on 02.03.17.
+ */
+public class WrongOptionsExeption extends RuntimeException {
+
+    public WrongOptionsExeption (String wrongOptionExeption){
+        super(wrongOptionExeption);
+
+    }
+
+}
