+++ src/main/java/peanut/medicine/newSurvey/Answer.java
@@ -1,0 +1,35 @@
+/**
+ * Created by bartman3000 on 2017-02-03.
+ */
+
+package peanut.medicine.newSurvey;
+
+public class Answer {
+    private Integer number;
+    private String text;
+    private String value;
+
+    public Integer getNumber() {
+        return number;
+    }
+
+    public void setNumber(Integer number) {
+        this.number = number;
+    }
+
+    public String getText() {
+        return text;
+    }
+
+    public void setText(String text) {
+        this.text = text;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+}
