+++ src/main/java/peanut/medicine/survey/SurveyPatient.java
@@ -1,0 +1,30 @@
+package peanut.medicine.survey;
+
+
+/**
+ * Created by moody on 20.02.17.
+ */
+public class SurveyPatient {
+
+
+ private String question;
+ private String[] answers ;
+
+    public String[] getAnswers() {
+        return answers;
+    }
+
+    public void setAnswers(String[] answers) {
+        this.answers = answers;
+    }
+
+    public String getQuestion() {
+        return question;
+    }
+
+    public void setQuestion(String question) {
+        this.question = question;
+    }
+
+
+}
