+++ src/main/java/peanut/medicine/survey/SurveyQuestionList.java
@@ -1,0 +1,17 @@
+package peanut.medicine.survey;
+
+/**
+ * Created by moody on 20.02.17.
+ */
+public class SurveyQuestionList {
+
+   private SurveyPatient[] surveyPatients ;
+
+    public SurveyPatient[] getSurveyPatients(){
+        return surveyPatients ;
+    }
+
+    public void setSurveyPatients(SurveyPatient[] surveyPatients) {
+        this.surveyPatients = surveyPatients;
+    }
+}
