+++ src/main/resources/survey.json
@@ -4,22 +4,19 @@
       "number" : 1,
       "name" : "name",
       "text" : "Imię",
-      "questionType" : "open",
-      "answerExpectedType" : "String"
+      "questionType" : "open"
     },
     {
       "number" : 2,
       "name" : "surname",
       "text" : "Nazwisko",
-      "questionType" : "open",
-      "answerExpectedType" : "String"
+      "questionType" : "open"
     },
     {
       "number" : 3,
       "name" : "sex",
       "text" : "Płeć",
       "questionType" : "closed",
-      "answerExpectedType" : "String",
       "answers": [
         {"number" : "1", "text" : "Kobieta", "value" : "woman"},
         {"number" : "2", "text" : "Mężczyzna", "value" : "man"}
@@ -29,15 +26,13 @@
       "number" : 4,
       "name" : "pesel",
       "text" : "Pesel",
-      "questionType" : "open",
-      "answerExpectedType" : "Integer"
+      "questionType" : "open"
     },
     {
       "number" : 5,
       "name" : "preferedSpecialization",
       "text" : "Jaka specjalizacja lekarza?",
       "questionType" : "closed",
-      "answerExpectedType" : "String",
       "answers": [
         {"number" : "1", "text" : "dentysta", "value" : "dentysta"},
         {"number" : "2", "text" : "internista", "value" : "internista"},
@@ -50,7 +45,6 @@
       "name" : "preferedDay",
       "text" : "Jaki dzień tygodnia?",
       "questionType" : "closed",
-      "answerExpectedType" : "String",
       "answers": [
         {"number" : "1", "text" : "poniedziałek", "value" : "monday"},
         {"number" : "2", "text" : "wtorek", "value" : "tuesday"},
